TalentaTalkBackend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
aiohttp
requests
numpy
pandas
librosa
torch
transformers
openai-whisper
google-generativeai
soundfile
scipy
pytz
openpyxl>=3.1.2
gruut>=2.2.3
xlsxwriter>=3.1.9